+++ src/app/api/alpaca/assets/route.ts
@@ -1,25 +1,17 @@
 import { authOptions } from '@/app/lib/auth';
+import { basicFetch } from '@/app/lib/fetchFunctions';
+import { AlpacaAssetModel } from '@/models/alpaca/alpaca-asset-model';
+import { ErrorCode } from '@/models/common/error-code';
 import { getServerSession } from 'next-auth';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(req: NextRequest) {
-
-    const session = await getServerSession(authOptions);  
-
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return NextResponse.json({ error: ErrorCode.Unauthorized });
+    }  
     const endpoint = `${process.env.ALPACA_API_URL}/AlpacaTrading/assets`;
-    const res = await fetch(endpoint, {
-      method: 'GET',
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${session!.accessToken}`,
-      }
-    });
-  
-    if (res.ok) {
-      const data = await res.json();
-  
-      return NextResponse.json(data);
-    } else {
-      return NextResponse.json({ error: res.statusText }, { status: res.status });
-    }
+    var dats = await basicFetch<AlpacaAssetModel[]>(endpoint, session.accessToken);
+    return NextResponse.json(dats);
+
 }