+++ src/app/api/strategy/list/route.ts
@@ -8,38 +8,16 @@ import { NextRequest, NextResponse } from "next/server";
 // TODO split this to 2 endpoints
 
 export async function GET(req: NextRequest) {
+
     const bookmarked = req.nextUrl.searchParams.get('bookmarked') as string;
-  
-        const session = await getServerSession(authOptions);
-        if (!session) {
-            return NextResponse.json({ error: ErrorCode.Unauthorized });
-        }
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json({ error: ErrorCode.Unauthorized });
+    }
 
-        const userId = session.user?.id;
-        if (!userId) {
-            return NextResponse.json({ error: ErrorCode.Unauthorized });
-        }
-        // console.log('session:', session);
-        // console.log('session accessToken :', session.accessToken);
-        // console.log('session jwt:', authOptions.jwt);
-
-        const endpoint = `${process.env.STRATEGY_API_URL}/strategy/settings/${userId}?bookmarked=${bookmarked}`;
-        const res = await fetch(endpoint, {
-            method: 'GET',
-            headers: {
-              "Content-Type": "application/json",
-              "Authorization": `Bearer ${session.accessToken}`,
-            }
-          });
-        
-          if (res.ok) {
-            const data = await res.json();
-        
-            return NextResponse.json(data);
-          } else {
-            return NextResponse.json({ error: res.statusText }, { status: res.status });
-          }
-   
+    const endpoint = `${process.env.STRATEGY_API_URL}/strategy/settings?bookmarked=${bookmarked}`;
+    var dats = await basicFetch<StrategySettings[]>(endpoint, session.accessToken);
+    return NextResponse.json(dats);
 }
 
 export async function PUT(req: NextRequest) {
@@ -88,9 +66,9 @@ export async function DELETE(req: NextRequest) {
             return NextResponse.json({ error: ErrorCode.Unauthorized });
         }
 
-        const testId = req.nextUrl.searchParams.get('testId') as string; 
+        const testId = req.nextUrl.searchParams.get('testId') as string;
         var endpoint = `${process.env.STRATEGY_API_URL}/strategy/${testId}`;
-    
+
         const options: RequestInit = {
             method: 'DELETE',
             headers: {
@@ -105,7 +83,7 @@ export async function DELETE(req: NextRequest) {
         const result = { message: message, success: success, errors: {} }
         return NextResponse.json(result);
 
-    } catch (error) {        
+    } catch (error) {
         return NextResponse.json({ error: 'Failed to fetch data' }, { status: 500 });
     }
 }