+++ src/app/api/strategy/test-positions/route.ts
@@ -1,25 +1,19 @@
 import { authOptions } from "@/app/lib/auth";
+import { basicFetch } from "@/app/lib/fetchFunctions";
+import { ErrorCode } from "@/models/common/error-code";
+import { PositionModel } from "@/models/strategy/position-model";
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(req: NextRequest) {
-    const testId = req.nextUrl.searchParams.get('testId') as string;
-    const session = await getServerSession(authOptions);
-
-    const endpoint = `${process.env.STRATEGY_API_URL}/strategy/positions/${testId}`;
-    const res = await fetch(endpoint, {
-        method: 'GET',
-        headers: {
-            "Content-Type": "application/json",
-            "Authorization": `Bearer ${session!.accessToken}`,
+    
+        const session = await getServerSession(authOptions);
+        if (!session) {
+            return NextResponse.json({ error: ErrorCode.Unauthorized });
         }
-    });
+        const testId = req.nextUrl.searchParams.get('testId') as string;
+        const endpoint = `${process.env.STRATEGY_API_URL}/strategy/positions/${testId}`;
 
-    if (res.ok) {
-        const data = await res.json();
-
-        return NextResponse.json(data);
-    } else {
-        return NextResponse.json({ error: res.statusText }, { status: res.status });
-    }
+        var dats = await basicFetch<PositionModel>(endpoint, session.accessToken);
+        return NextResponse.json(dats);
 }